Update README.md
This repository delves into the application of sentiment analysis in evaluating customer feedback for products, with a specific focus on e-commerce platforms such as Amazon. It encompasses the utilization of natural language processing (NLP) and machine learning methodologies to categorize reviews as either positive or negative, accomplishing accuracies of a minimum of 80%. The research underscores the formidable challenge posed by the substantial volume of daily product reviews on online marketplaces and underscores the value of sentiment analysis in comprehending customer perspectives. The study asserts that this approach can effectively aid both consumers and product manufacturers in making informed decisions.